backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "swarlipi-backend"
version = "0.1.0"
description = "SwarLipi AI - AI-Powered Music Transcription Platform Backend"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "websockets>=12.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.9",
    # Task Queue
    "celery[redis]>=5.4.0",
    "redis>=5.0.0",
    # AI / ML
    "torch>=2.2.0",
    "torchaudio>=2.2.0",
    "music21>=9.1.0",
    "pretty-midi>=0.2.10",
    "librosa>=0.10.1",
    "soundfile>=0.12.1",
    "numpy>=1.26.0",
    # Cloud Storage
    "google-cloud-storage>=2.16.0",
    # Audio Processing
    "yt-dlp>=2024.3.10",
    "pydub>=0.25.1",
    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    # Utilities
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "aiofiles>=23.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.3.0",
    "black>=24.3.0",
    "mypy>=1.9.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "COM", "C4", "SIM"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
